Added cmd options and stats

+++ include/klee/ExecutionState.h
@@ -146,7 +146,7 @@ public:
   void removeFnAlias(std::string fn);
   
   /// @brief Bytes of memory currently allocated for this state (and it's predecessors)
-  uint64_t memoryUsage;
+  unsigned memoryUsage;
 
 private:
   ExecutionState() : ptreeNode(0) {}
